requirements.txt

 
pytz==2023.3
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
httpx==0.27.0
aiomysql==0.2.0
alembic==1.13.2
ruff==0.5.5
pytest==8.2.2
python-multipart==0.0.9
boto3==1.34.160
firebase-admin==6.5.0
